Feature/warnings compiler flags

See merge request CAMP_IFL/simple!33

Former-commit-id: 4fe7ba2cfb94cc30e3b300e4460ac4919499067d

+++ examples/publisher.cpp
@@ -27,7 +27,7 @@ using namespace std::chrono;
 double getTimeNow()
 {
   auto now = system_clock::now();
-  return time_point_cast<milliseconds>(now).time_since_epoch().count();
+  return static_cast<double>(time_point_cast<milliseconds>(now).time_since_epoch().count());
 }
 
 int main()
